refactored start sync into a separate file, implemented local unison launch and checking for readiness
